+++ src/types/ErrorTypes.ts
@@ -3,11 +3,13 @@ export type PlanErrorLocation =
   | 'account_initial_login_code'
   | 'account_get_plans'
   | 'account_create_plan'
+  | 'account_edit_plan'
   | 'account_delete_plan'
   | 'account_activate_plan'
   | 'account_update_plan'
   | 'account_get_schedules'
   | 'account_create_schedule'
+  | 'account_edit_schedule'
   | 'account_delete_schedule'
   | 'account_activate_schedule'
   | 'account_update_schedule';